test: split the promoted media scenarios into focused modules
Both suites had grown past 1,500 lines with the promoted-widget scenarios
sharing fixture state with everything else. Move them out, following the
existing <module>.<topic>.test.ts convention:

  useErrorClearingHooks.test.ts   1,658 -> 1,371  (+ .promotion.test.ts)
  missingMediaScan.test.ts        1,513 -> 1,225  (+ .promotedWidget.test.ts)

Test bodies are unchanged and the counts match on both sides of the move.
